여러가지
앉았다 일어났을 때 생기는 버그 수정
던전 룸 만드는 중...
신호등-구슬 눌러 문 활성화 구현

+++ Team Project E/Assets/Scripts/PlayerUI/PlayerstatManagement_L.cs
@@ -38,7 +38,7 @@ public class PlayerstatManagement_L : MonoBehaviour
     public void HPtext()
     {
         float TextHP = myPlayer.GetComponent<SPlayer>().MyStatus.HP;
-        playerHP.text = TextHP + " / " + "100";
+        playerHP.text = (int)TextHP + " / " + "100";
         HPSlider.value = TextHP / 100;
 
     }
